Move CoreService to production
