+++ Sources/Curtz/Core Service/ShortenResult.swift
@@ -1,0 +1,13 @@
+//
+//  ShortenResult.swift
+//  Curtz
+//
+//  Created by George Nyakundi on 20/11/2023.
+//
+
+import Foundation
+
+public enum ShortenResult {
+    case success(ShortenResponseItem)
+    case failure(Error)
+}
