+++ Tests/Curtz/Core/CoreServiceUnitTests.swift
@@ -9,178 +9,7 @@ import XCTest
 import Curtz
 
 
-final class CoreServiceResponseMapper {
-    
-    private struct ShortenItem: Decodable {
-        let id: String
-        let custom_alias: String
-        let original_url: String
-        let expires_on: String
-        let keywords: [String]
-        let user_id: String
-        let short_code: String
-        let created_at: String
-        let updated_at: String
-        let hits: Int
-        
-        var response: ShortenResponseItem {
-            return ShortenResponseItem(
-                id: id,
-                customAlias: custom_alias,
-                originalUrl: original_url,
-                expiresOn: expires_on,
-                keywords: keywords,
-                userId: user_id,
-                shortCode: short_code,
-                createdAt: created_at,
-                updatedAt: updated_at,
-                hits: hits
-            )
-        }
-    }
-    
-    private struct ErrorItem: Decodable {
-        let error: String
-    }
-    
-    private static var decoder: JSONDecoder = {
-        let decoder = JSONDecoder()
-        decoder.dateDecodingStrategy = .iso8601
-        return decoder
-    }()
-    
-    
-    
-    static func mapShorteningResponse(_ data: Data, from response: HTTPURLResponse) -> CoreService.ShorteningResult {
-        
-        if response.isBadRequest() {
-            let res = try? decoder.decode(ErrorItem.self, from: data)
-            return .failure(CoreService.Error.clientError(res?.error ?? ""))
-        }
-        
-        guard response.isOK(), let res = try? decoder.decode(ShortenItem.self, from: data) else {
-            return .failure(CoreService.Error.invalidResponse)
-        }
-        
-        return .success(res.response)
-    }
-    
-    static func mapFetchAllResponse(_ data: Data, from response: HTTPURLResponse) -> CoreService.FetchResult {
-        if response.isBadRequest() {
-            let res = try? decoder.decode(ErrorItem.self, from: data)
-            return .failure(CoreService.Error.clientError(res?.error ?? ""))
-        }
-        
-        guard response.isOK(), let res = try?
-                decoder.decode([ShortenItem].self, from: data) else {
-            return .failure(CoreService.Error.invalidResponse)
-        }
-        
-        return .success(res.map { $0.response })
-    }
-}
 
-struct ShortenRequest {
-    let originalUrl: String
-    let customAlias: String
-    let keywords: [String]
-    let expiresOn: String
-    
-    init(originalUrl: String, customAlias: String, keywords: [String], expiresOn: String) {
-        self.originalUrl = originalUrl
-        self.customAlias = customAlias
-        self.keywords = keywords
-        self.expiresOn = expiresOn
-    }
-}
-
-struct ShortenResponseItem {
-    let id: String
-    let customAlias: String
-    let originalUrl: String
-    let expiresOn: String
-    let keywords: [String]
-    let userId: String
-    let shortCode: String
-    let createdAt: String
-    let updatedAt: String
-    let hits: Int
-    
-    init(id: String, customAlias: String, originalUrl: String, expiresOn: String, keywords: [String], userId: String, shortCode: String, createdAt: String, updatedAt: String, hits: Int) {
-        self.id = id
-        self.customAlias = customAlias
-        self.originalUrl = originalUrl
-        self.expiresOn = expiresOn
-        self.keywords = keywords
-        self.userId = userId
-        self.shortCode = shortCode
-        self.createdAt = createdAt
-        self.updatedAt = updatedAt
-        self.hits = hits
-    }
-}
-
-extension ShortenResponseItem: Equatable {}
-
-enum ShortenResult {
-    case success(ShortenResponseItem)
-    case failure(Error)
-}
-
-extension CoreService.Error: Equatable { }
-
-class CoreService {
-    private let client: HTTPClient
-    private let serviceURL: URL
-    
-    typealias ShorteningResult = ShortenResult
-    typealias FetchResult = Result<[ShortenResponseItem], Error>
-    
-    enum Error: Swift.Error {
-        case malformedRequest
-        case invalidResponse
-        case clientError(String)
-    }
-    
-    init(serviceURL: URL, client: HTTPClient) {
-        self.serviceURL = serviceURL
-        self.client = client
-    }
-    
-    func shorten(_ request: ShortenRequest, completion: @escaping(ShorteningResult) -> Void) {
-        client.perform(
-            request: .prepared(
-                for: .shortening(
-                    originalUrl: request.originalUrl,
-                    customAlias: request.customAlias,
-                    keywords: request.keywords,
-                    expiresOn: request.expiresOn
-                ),
-                with: serviceURL
-            )
-        ) { [weak self] result in
-            guard self != nil else { return }
-            switch result {
-            case let .success((data, response)):
-                completion(CoreServiceResponseMapper.mapShorteningResponse(data, from: response))
-            default:
-                completion(.failure(Error.malformedRequest))
-            }
-        }
-    }
-    
-    func fetchAll(completion: @escaping(FetchResult) -> Void){
-        client.perform(request: .prepared(for: .fetching, with: serviceURL)) {[weak self] result  in
-            guard self != nil else { return }
-            switch result {
-            case let .success((data, response)):
-                completion(CoreServiceResponseMapper.mapFetchAllResponse(data, from: response))
-            case .failure:
-                completion(.failure(.invalidResponse))
-            }
-        }
-    }
-}
 
 final class CoreServiceUnitTests: XCTestCase {
     
@@ -393,7 +222,7 @@ final class CoreServiceUnitTests: XCTestCase {
         ShortenRequest(originalUrl: "https://sampleUrl.com",customAlias: "alias", keywords: [], expiresOn: "")
     }
     
-    private struct TestShortenRequest: Decodable {
+    struct TestShortenRequest: Decodable {
         let originalUrl: String
         let customAlias: String
         let keywords: [String]
